requirements.txt
Django>=4.2
mysqlclient>=2.2.0
djangorestframework
drf-yasg
pytest
pytest-django
